backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
argon2-cffi==23.1.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Services
# torch and torchvision - installed separately via Dockerfile for CPU-only
git+https://github.com/openai/whisper.git
google-generativeai==0.3.2
easyocr==1.7.1
ffmpeg-python==0.2.0

# Image Processing
numpy<2.0.0  # Fix compatibility with OpenCV and EasyOCR
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Cloud Storage - Not needed for local development
# boto3==1.34.34
# botocore==1.34.34

# Utilities
python-dateutil==2.8.2
pytz==2024.1
httpx==0.26.0
tenacity==8.2.3

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing (optional, for development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality (optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0
